cleaner division of logic
